260616
- 피격 vfx 교체
- 천둥 구름 스킬 수정 및 vfx 교체
- 스킬 이미지 교체
- 스킬 string data 정리
- 스킬 데이터 및 프리팹 수정

+++ Assets/ProjectUD/Scripts/Effect/FearEffect.cs
@@ -6,7 +6,7 @@ public class FearEffect : DurationEffect
 
     [Header("■ Fear Options")]
     [SerializeField] private float atkPercent;
-    [SerializeField] private float mental;
+    [SerializeField] private int mental;
 
     public override void Activate()
     {
